+++ apps/examples/dpp/philo.c
@@ -170,7 +170,7 @@ void philo_create(
     me->cnt = 0;
     me->id = id;
     am_ao_create(&me->ao, (am_ao_fn)am_hsm_init, (am_ao_fn)am_hsm_dispatch, me);
-    am_hsm_create(&me->hsm, am_hsm_state(philo_init));
+    am_hsm_create(&me->hsm, am_hsm_state_make(philo_init));
 
     me->table = table;
     me->timer = timer;