+++ apps/examples/watchdog/main.c
@@ -108,7 +108,7 @@ static void watched_create(
 ) {
     memset(me, 0, sizeof(*me));
     am_ao_create(&me->ao, (am_ao_fn)am_hsm_init, (am_ao_fn)am_hsm_dispatch, me);
-    am_hsm_create(&me->hsm, am_hsm_state(watched_init));
+    am_hsm_create(&me->hsm, am_hsm_state_make(watched_init));
     me->timer = timer;
     me->feed = am_timer_event_create_x(EVT_WATCHED_TIMEOUT, &me->ao);
     me->wdt = wdt;
@@ -147,7 +147,7 @@ static enum am_rc wdt_init(struct am_hsm* hsm, const struct am_event* event) {
 static void wdt_create(struct wdt* me, struct am_timer* timer) {
     memset(me, 0, sizeof(*me));
     am_ao_create(&me->ao, (am_ao_fn)am_hsm_init, (am_ao_fn)am_hsm_dispatch, me);
-    am_hsm_create(&me->hsm, am_hsm_state(wdt_init));
+    am_hsm_create(&me->hsm, am_hsm_state_make(wdt_init));
     me->timer = timer;
     me->bark = am_timer_event_create_x(EVT_WDT_BARK, &me->ao);
 }