+++ libs/hsm/tests/redispatch.c
@@ -85,7 +85,7 @@ static enum am_rc redisp_sinit(
 
 static void test_redispatch(void) {
     struct test_redisp* me = &m_test_redisp;
-    am_hsm_create(&me->hsm, am_hsm_state(redisp_sinit));
+    am_hsm_create(&me->hsm, am_hsm_state_make(redisp_sinit));
 
     am_hsm_init(&me->hsm, /*init_event=*/NULL);
     AM_ASSERT(0 == me->foo);
@@ -93,12 +93,12 @@ static void test_redispatch(void) {
     static const struct am_event e1 = {.id = HSM_EVT_A};
     am_hsm_dispatch(&me->hsm, &e1);
     AM_ASSERT(1 == me->foo);
-    AM_ASSERT(am_hsm_state_is_eq(&me->hsm, am_hsm_state(redisp_s2)));
+    AM_ASSERT(am_hsm_state_is_eq(&me->hsm, am_hsm_state_make(redisp_s2)));
 
     static const struct am_event e2 = {.id = HSM_EVT_B};
     am_hsm_dispatch(&me->hsm, &e2);
     AM_ASSERT(2 == me->foo2);
-    AM_ASSERT(am_hsm_state_is_eq(&me->hsm, am_hsm_state(redisp_s1)));
+    AM_ASSERT(am_hsm_state_is_eq(&me->hsm, am_hsm_state_make(redisp_s1)));
 }
 
 int main(void) {