+++ libs/hsm/tests/regular/regular.c
@@ -70,7 +70,7 @@ static enum am_rc r_s(struct am_hsm* hsm, const struct am_event* event) {
     switch (event->id) {
     case AM_EVT_ENTRY: {
         me->log("s-ENTRY;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
 
         struct am_hsm_state state = am_hsm_get_state(hsm);
         AM_ASSERT(am_hsm_state_is_eq(hsm, state));
@@ -79,7 +79,7 @@ static enum am_rc r_s(struct am_hsm* hsm, const struct am_event* event) {
     }
     case AM_EVT_INIT: {
         me->log("s-INIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
 
         struct am_hsm_state state = am_hsm_get_state(hsm);
         AM_ASSERT(am_hsm_state_is_eq(hsm, state));
@@ -88,7 +88,7 @@ static enum am_rc r_s(struct am_hsm* hsm, const struct am_event* event) {
     }
     case AM_EVT_EXIT: {
         me->log("s-EXIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
 
         struct am_hsm_state state = am_hsm_get_state(hsm);
         AM_ASSERT(am_hsm_state_is_eq(hsm, state));
@@ -96,9 +96,9 @@ static enum am_rc r_s(struct am_hsm* hsm, const struct am_event* event) {
         return am_hsm_handled(hsm);
     }
     case HSM_EVT_I: {
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(am_hsm_top)));
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s211)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(am_hsm_top)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s211)));
         if (me->foo) {
             me->foo = 0;
             me->log("s-I;");
@@ -108,8 +108,8 @@ static enum am_rc r_s(struct am_hsm* hsm, const struct am_event* event) {
     }
     case HSM_EVT_E:
         AM_ASSERT(
-            (am_hsm_state_is_eq(hsm, am_hsm_state(r_s211))) ||
-            (am_hsm_state_is_eq(hsm, am_hsm_state(r_s11)))
+            (am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s211))) ||
+            (am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s11)))
         );
         me->foo = 0;
         me->log("s-E;");
@@ -131,35 +131,35 @@ static enum am_rc r_s1(struct am_hsm* hsm, const struct am_event* event) {
     switch (event->id) {
     case AM_EVT_ENTRY: {
         me->log("s1-ENTRY;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
         struct am_hsm_state state = am_hsm_get_state(hsm);
         AM_ASSERT(am_hsm_state_is_eq(hsm, state));
         return am_hsm_handled(hsm);
     }
     case AM_EVT_INIT: {
         me->log("s1-INIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
         struct am_hsm_state state = am_hsm_get_state(hsm);
         AM_ASSERT(am_hsm_state_is_eq(hsm, state));
         return am_hsm_tran(hsm, r_s11);
     }
     case AM_EVT_EXIT: {
         me->log("s1-EXIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
         struct am_hsm_state state = am_hsm_get_state(hsm);
         AM_ASSERT(am_hsm_state_is_eq(hsm, state));
         return am_hsm_handled(hsm);
     }
     case HSM_EVT_I:
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s1)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(am_hsm_top)));
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(am_hsm_top)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s11)));
         me->log("s1-I;");
         return am_hsm_handled(hsm);
 
     case HSM_EVT_C:
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s11)));
         me->log("s1-C;");
         return am_hsm_tran(hsm, r_s2);
 
@@ -169,7 +169,7 @@ static enum am_rc r_s1(struct am_hsm* hsm, const struct am_event* event) {
 
     case HSM_EVT_A:
         me->log("s1-A;");
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s11)));
         return am_hsm_tran(hsm, r_s1);
 
     case HSM_EVT_B:
@@ -177,7 +177,7 @@ static enum am_rc r_s1(struct am_hsm* hsm, const struct am_event* event) {
         return am_hsm_tran(hsm, r_s11);
 
     case HSM_EVT_D:
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s11)));
         if (!me->foo) {
             me->foo = 1;
             me->log("s1-D;");
@@ -197,25 +197,25 @@ static enum am_rc r_s11(struct am_hsm* hsm, const struct am_event* event) {
     switch (event->id) {
     case AM_EVT_ENTRY:
         me->log("s11-ENTRY;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s11)));
         return am_hsm_handled(hsm);
 
     case AM_EVT_EXIT:
         me->log("s11-EXIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s11)));
         return am_hsm_handled(hsm);
 
     case AM_EVT_INIT:
         me->log("s11-INIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s11)));
         return am_hsm_handled(hsm);
 
     case HSM_EVT_G:
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s11)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s1)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(am_hsm_top)));
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s11)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(am_hsm_top)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s11)));
         me->log("s11-G;");
         return am_hsm_tran(hsm, r_s211);
 
@@ -224,7 +224,7 @@ static enum am_rc r_s11(struct am_hsm* hsm, const struct am_event* event) {
         return am_hsm_tran(hsm, r_s);
 
     case HSM_EVT_D:
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s11)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s11)));
         if (me->foo) {
             me->foo = 0;
             me->log("s11-D;");
@@ -244,26 +244,26 @@ static enum am_rc r_s2(struct am_hsm* hsm, const struct am_event* event) {
     switch (event->id) {
     case AM_EVT_ENTRY:
         me->log("s2-ENTRY;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s2)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s2)));
         return am_hsm_handled(hsm);
 
     case AM_EVT_INIT:
         me->log("s2-INIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s2)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s2)));
         return am_hsm_tran(hsm, r_s211);
 
     case AM_EVT_EXIT:
         me->log("s2-EXIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s2)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s2)));
         return am_hsm_handled(hsm);
 
     case HSM_EVT_I:
-        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state(r_s11)));
-        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state(r_s1)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s2)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(am_hsm_top)));
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s211)));
+        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state_make(r_s11)));
+        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s2)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(am_hsm_top)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s211)));
         if (!me->foo) {
             me->foo = 1;
             me->log("s2-I;");
@@ -291,28 +291,28 @@ static enum am_rc r_s21(struct am_hsm* hsm, const struct am_event* event) {
     switch (event->id) {
     case AM_EVT_ENTRY:
         me->log("s21-ENTRY;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s21)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s21)));
         return am_hsm_handled(hsm);
 
     case AM_EVT_INIT:
         me->log("s21-INIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s21)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s21)));
         return am_hsm_tran(hsm, r_s211);
 
     case AM_EVT_EXIT:
         me->log("s21-EXIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s21)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s21)));
         return am_hsm_handled(hsm);
 
     case HSM_EVT_A: {
         me->log("s21-A;");
-        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state(r_s11)));
-        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state(r_s1)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s21)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s2)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(am_hsm_top)));
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s211)));
+        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state_make(r_s11)));
+        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s21)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s2)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(am_hsm_top)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s211)));
 
         struct am_hsm_state state = am_hsm_get_state(hsm);
         AM_ASSERT(am_hsm_state_is_eq(hsm, state));
@@ -325,11 +325,11 @@ static enum am_rc r_s21(struct am_hsm* hsm, const struct am_event* event) {
 
     case HSM_EVT_G:
         me->log("s21-G;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s21)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s2)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(am_hsm_top)));
-        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state(r_s211)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s21)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s2)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(am_hsm_top)));
+        AM_ASSERT(am_hsm_state_is_eq(hsm, am_hsm_state_make(r_s211)));
         return am_hsm_tran(hsm, r_s1);
 
     default:
@@ -344,27 +344,27 @@ static enum am_rc r_s211(struct am_hsm* hsm, const struct am_event* event) {
     switch (event->id) {
     case AM_EVT_ENTRY:
         me->log("s211-ENTRY;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s211)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s211)));
         return am_hsm_handled(hsm);
 
     case AM_EVT_EXIT:
         me->log("s211-EXIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s211)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s211)));
         return am_hsm_handled(hsm);
 
     case AM_EVT_INIT:
         me->log("s211-INIT;");
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s211)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s211)));
         return am_hsm_handled(hsm);
 
     case HSM_EVT_D:
-        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state(r_s11)));
-        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state(r_s1)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s211)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s21)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s2)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(r_s)));
-        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state(am_hsm_top)));
+        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state_make(r_s11)));
+        AM_ASSERT(!am_hsm_is_in(hsm, am_hsm_state_make(r_s1)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s211)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s21)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s2)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(r_s)));
+        AM_ASSERT(am_hsm_is_in(hsm, am_hsm_state_make(am_hsm_top)));
         me->log("s211-D;");
         return am_hsm_tran(hsm, r_s21);
 
@@ -380,7 +380,7 @@ static enum am_rc r_s211(struct am_hsm* hsm, const struct am_event* event) {
 
 void regular_create(void (*log)(const char* fmt, ...)) {
     struct regular* me = &m_regular;
-    am_hsm_create(&me->hsm, am_hsm_state(regular_init));
+    am_hsm_create(&me->hsm, am_hsm_state_make(regular_init));
     me->log = log;
 }