+++ CHANGELOG.md
@@ -94,6 +94,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
   async functions.
 - Add `struct am_ao::running` as a safety guard against using an uninitialized
   active object
+- Rename `am_hsm_state()` to `am_hsm_state_make()`
 
 ### Fixed